fix: add instruction for running coverage tool to complement configuration (#119)
This PR adds an important instruction about `--coverage` flag that jest expects. Jest doesn't document it because an application developer is expected to pass this flag. Istanbul doesn't have it because we talk about configuration. This tiny detail gets lost between two worlds and so takes longer for people like me to configure istanbul and actual trigger a run.